Remove check for out of order interface
The interfaces are returned with out of order index when they are hot
swapped(e.g., USB Ethernet Adapter). I don't understand why the
condition needs to be checked in the first place as we don't use the
index anywhere.

Reported the issue here https://github.com/google/gopacket/issues/830
but not response. So, removing it to make it work in hot swapped network
interfaces.