Fix out of order interface (#3)
The interfaces are returned with out of order index when they are hot
swapped(e.g., USB Ethernet Adapter). Couldn't understand why the
condition needs to be checked in the first place as the code doesn't use the
index anywhere.

Reported the issue here https://github.com/google/gopacket/issues/830
but not response. So, removing it to make it work in hot swapped network
interfaces.